requirements.txt
asgiref==3.3.2
Django==3.2.8
djangorestframework==3.12.2
django-cors-headers==3.7.0
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.9.1
djangorestframework==3.12.2
django-cors-headers==3.7.0